Avoid try/catch for when reading keys of built-ins in util.inspect
The `getKeys` function currently needs to wrap the `ObjectKeys` call
in a try/catch statement because, if the object is a module namespace,
it might throw.

This commit avoids needing `try/catch` by not calling `getKeys` on
module namespace objects. There was an old comment about not
wanting to perform this branching because it's expensive, but all
the calls to `getKeys` are already brand-checked so the caller knows.